Fix manual input on Firefox when maxRange is set (fixes #309)
This would not work in Firefox when the whole value was selected, since
the code would incorrectly check the value-length on Firefox on
keypresses. This did not cause problems in Chrome, since there the
selection is available for number inputs and the value-length condition
would not be checked at all.